Update schema, refactoring

+++ src/main/java/com/app/recipes/entity/Recipe.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 
-import java.time.Instant;
 import java.time.LocalDateTime;
 
 @Getter
@@ -16,7 +15,7 @@ public class Recipe {
     @Id
     @ColumnDefault("nextval('recipe_id_seq')")
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", nullable = false)
     private String name;
